React better to ejecting of CDs during ripping
cdio API is really bad.
This makes the program exit with an error rather than hang or
spam a wall of messages or worse.
More importantly, could you not remove CDs during ripping?
There's an API to lock the drive door, but its very prone to desyncs,
so it could become stuck until an actual system reboot. And its also
very much not standard.