Mark version 0.4.0-rc
Premliminary changelog:
    Full pregap handling
    Tray doesn't eject by default
    Fix big endian systems
    Speed up by reducing memory copies
    Skip reading data tracks

You may want to check your entire collection for pregaps and whether
the way they used to be handled is still okay with you.

This patch was brought to you by Opeth's Ghost Reveries album, which one day.
I hope to rip the way it must be - Reverie split off or merged into the next track.