Release version 0.4.2
Changelog:
    Full pregap handling
    HDCD decoding
    One log per output folder
    Logging format improvements
    Continue if manual metadata is specified and MB meta is missing
    Better MusicBrainz error reporting
    Tray doesn't eject by default
    Fix theoretical big endian systems
    Speed up by reducing memory copies
    Ignore reading data tracks
    More robust error detection

I don't know what else to improve, so this will be the version to use
for a while. This is now perfectly usable for my needs.

Better error correction would be nice, but cd-paranoia is already very good,
and none of the incredibly scratched discs have made it fail or error.