Release 0.4.3 - Multi-disc ripping
Okay, *this* is the last feature I needed.
And my CD drive is badly damaged now so this should be the last version
in a while. Byebye LITE-ON iHAS124 F. You were noisy, and you've gotten
hypernoisy and your tray is hard to close.
discnumber is the standard tag, but libavformat will convert disc to
discnumber for all formats. totaldiscs seems largely unknown, but whatever.

Also I wish musicbrainz would also provide the disc number but it doesn't.