Verify partially damaged tracks via accurip
AccuRip sends you the __CHECKSUMS__ of the 450th frame,
probably with the intent to use them to find the drive
offset. Which we do.

But we can also use them to verify only a single frame of
a track. Its not much, but having this match while
v1 and v2 mismatch means the disc or drive are somewhat damaged.
So we only print this if there's no AccuRip entry found,
or the other 2 don't match.

This feature doesn't exist in any other ripping program.