Release version 0.7.0
This project is now feature complete. It does everything I'd like for it to do
in order to rip my entire CD collection. There's still some polishing work
to optionally do, like read the JSON directly from MBDB and CADB, as it
carries more data like release events, and the cover art DB does contain
some more images that we don't fetch yet. And no doubt there are a few bugs
left to fix, like making PSX game CDs more accurately read.

This project was originally meant to be also usable as a library for embedding,
so maybe if there's some interest to do this work it wouldn't take long to do;
the project doesn't have too much bad code and hacks, and the main header
isn't as messy as it could have otherwise been.

Thanks to all contributors and testers.