musicbrainz: do not append all tags by default
The way the awful credits musicbrainz API works is that you need
to append each credit with a joiner string. The common macro
we used to do that did this for all strings. Which worked, as
long as the metadata field was completely empty. Which got changed.

So, instead of always appending, just append where we need to.