data tracks: rip in unprocessed 2352 byte mode
If you ask for autodetect or data sector type, the drive tries to verify
the ECC for you, even if you ask it not to. This means sectors with ECC
errors due to mastering can't be archived.

By reading as audio and fixing the offset up ourselves, we can avoid the
issue.