Fix segfault in cleanup.
This fixes a potential segfault when `avio_open` fails -> we go to
cleanup -> cleanup sees that `s->mutex_held` (now `s->mutex_status`) is true, and thus tries
to acquire the mutex, yet the mutex isn't initialized: segfault.

A similar issue occurs when we try to `pthread_join` on the thread in
the cleanup if it's not created. Thus a "bool" for the initialization of
the thread is added.

One condition that `avio_open` can fail on is when the file path is too
long.